+++ src/main/java/com/shawncurrie/techblogs/service/CompanyService.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface CompanyService {
     List<CompanyDTO> getCompanies(int page, int limit);
-    CompanyDTO getCompany(long companyId);
+    CompanyDTO getCompany(int companyId);
 }
